refactor: consolidate SiteRequestCompleted into RequestCompleted with optional deployment
