+++ src/Appwrite/Bus/Listeners/Usage.php
@@ -2,12 +2,9 @@
 
 namespace Appwrite\Bus\Listeners;
 
-use Appwrite\Bus\ExecutionCompleted;
-use Appwrite\Bus\RequestCompleted;
-use Appwrite\Bus\SiteRequestCompleted;
+use Appwrite\Bus\Events\ExecutionCompleted;
+use Appwrite\Bus\Events\RequestCompleted;
 use Appwrite\Event\StatsUsage;
-use Appwrite\Utopia\Request;
-use Appwrite\Utopia\Response;
 use Utopia\Bus\Event;
 use Utopia\Bus\Listener;
 use Utopia\Database\Document;
@@ -25,7 +22,6 @@ class Usage extends Listener
         return [
             ExecutionCompleted::class,
             RequestCompleted::class,
-            SiteRequestCompleted::class,
         ];
     }
 
@@ -41,7 +37,6 @@ class Usage extends Listener
     {
         match (true) {
             $event instanceof ExecutionCompleted => $this->handleExecutionCompleted($event, $publisher),
-            $event instanceof SiteRequestCompleted => $this->handleSiteRequestCompleted($event, $publisher),
             $event instanceof RequestCompleted => $this->handleRequestCompleted($event, $publisher),
             default => null,
         };
@@ -89,32 +84,31 @@ class Usage extends Listener
             $fileSize = (\is_array($file['size']) && isset($file['size'][0])) ? $file['size'][0] : $file['size'];
         }
 
+        $project = new Document($event->project);
+        $deployment = new Document($event->deployment);
         $queueForStatsUsage = new StatsUsage($publisher);
-        $queueForStatsUsage
-            ->setProject(new Document($event->project))
-            ->addMetric(METRIC_NETWORK_REQUESTS, 1)
-            ->addMetric(METRIC_NETWORK_INBOUND, $event->request->getSize() + $fileSize)
-            ->addMetric(METRIC_NETWORK_OUTBOUND, $event->response->getSize())
-            ->trigger();
-    }
 
-    private function handleSiteRequestCompleted(SiteRequestCompleted $event, Publisher $publisher): void
-    {
-        $fileSize = 0;
-        $file = $event->request->getFiles('file');
-        if (!empty($file)) {
-            $fileSize = (\is_array($file['size']) && isset($file['size'][0])) ? $file['size'][0] : $file['size'];
+        $inbound = $event->request->getSize() + $fileSize;
+        $outbound = $event->response->getSize();
+
+        $queueForStatsUsage->setProject($project);
+
+        if ($deployment->getAttribute('resourceType') === 'sites') {
+            $siteInternalId = $deployment->getAttribute('resourceInternalId', '');
+            $queueForStatsUsage
+                ->addMetric(METRIC_SITES_REQUESTS, 1)
+                ->addMetric(METRIC_SITES_INBOUND, $inbound)
+                ->addMetric(METRIC_SITES_OUTBOUND, $outbound)
+                ->addMetric(str_replace('{siteInternalId}', $siteInternalId, METRIC_SITES_ID_REQUESTS), 1)
+                ->addMetric(str_replace('{siteInternalId}', $siteInternalId, METRIC_SITES_ID_INBOUND), $inbound)
+                ->addMetric(str_replace('{siteInternalId}', $siteInternalId, METRIC_SITES_ID_OUTBOUND), $outbound);
+        } else {
+            $queueForStatsUsage
+                ->addMetric(METRIC_NETWORK_REQUESTS, 1)
+                ->addMetric(METRIC_NETWORK_INBOUND, $inbound)
+                ->addMetric(METRIC_NETWORK_OUTBOUND, $outbound);
         }
 
-        $queueForStatsUsage = new StatsUsage($publisher);
-        $queueForStatsUsage
-            ->setProject(new Document($event->project))
-            ->addMetric(METRIC_SITES_REQUESTS, 1)
-            ->addMetric(METRIC_SITES_INBOUND, $event->request->getSize() + $fileSize)
-            ->addMetric(METRIC_SITES_OUTBOUND, $event->response->getSize())
-            ->addMetric(str_replace('{siteInternalId}', $event->siteInternalId, METRIC_SITES_ID_REQUESTS), 1)
-            ->addMetric(str_replace('{siteInternalId}', $event->siteInternalId, METRIC_SITES_ID_INBOUND), $event->request->getSize() + $fileSize)
-            ->addMetric(str_replace('{siteInternalId}', $event->siteInternalId, METRIC_SITES_ID_OUTBOUND), $event->response->getSize())
-            ->trigger();
+        $queueForStatsUsage->trigger();
     }
 }
